started to port segments to a mixin class
