+++ cmsplugin_cascade/templates/cascade/generic/wrapper.html
@@ -1,6 +1,6 @@
 {% load cms_tags %}
-{% with tag_type=instance.tag_type css_classes=instance.css_classes inline_styles=instance.inline_styles %}
-<{{ tag_type }} {{ instance.html_tag_attributes }}{% if css_classes %} class="{{ css_classes }}"{% endif %}{% if inline_styles %} style="{{ inline_styles }}"{% endif %}>
+{% with tag_type=instance.tag_type css_classes=instance.css_classes inline_styles=instance.inline_styles element_id=glossary.element_id %}
+<{{ tag_type }} {{ instance.html_tag_attributes }}{% if element_id %} id="{{ element_id }}"{% endif %}{% if css_classes %} class="{{ css_classes }}"{% endif %}{% if inline_styles %} style="{{ inline_styles }}"{% endif %}>
 {% for plugin in instance.child_plugin_instances %}
 	{% render_plugin plugin %}
 {% endfor %}
